+++ ChangeLog.md
@@ -2,7 +2,7 @@
 
 ## 0.8.8.0  -- 2021-06-13
 
-* Add `withCPtrForever` and `mkCPtrDep` to `Z.Foreign.CPtr` module.
+* Add `withCPtrForever` and `addCPtrDep` to `Z.Foreign.CPtr` module.
 * Add `replicateMVec` to `Z.Data.Vector`.
 * Fix AVX512 text validate header building with C11.